restructure project

+++ lib/nes/controller.cc
@@ -7,7 +7,7 @@ namespace nes
 		auto const res = index_ < 8 ? (pressed_.get_raw_value() >> index_) & 1 : 0;
 		index_ += 1;
 		if (strobing_) { index_ = 0; }
-		return res;
+		return static_cast<std::uint8_t>(res);
 	}
 
 	auto controller::write(std::uint8_t const value) -> void
